requirements.txt
absl-py==2.4.0
attrs==25.4.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.12.19
fonttools==4.61.1
idna==3.11
ImageIO==2.37.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
lazy_loader==0.4
llvmlite==0.46.0
matplotlib==3.10.8
mediapipe==0.10.9
mpmath==1.3.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
onnxruntime==1.24.2
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
platformdirs==4.9.2
pooch==1.9.0
protobuf==3.20.3
pycparser==3.0
PyMatting==1.1.15
pyparsing==3.3.2
python-dateutil==2.9.0.post0
referencing==0.37.0
rembg==2.0.72
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
sounddevice==0.5.5
sympy==1.14.0
tifffile==2026.2.20
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
